+++ jest.config.ts
@@ -24,6 +24,8 @@ const config: Config.InitialOptions = {
       "<rootDir>/packages/implementations/react/src/atomic/IconButton/src",
     "@rarui-react/icon":
       "<rootDir>/packages/implementations/react/src/atomic/Icon/src",
+    "@rarui-react/divider":
+      "<rootDir>/packages/implementations/react/src/atomic/Divider/src",
     // composites
   },
   transform: {