+++ packages/implementations/react/src/atomic/Divider/src/index.ts
@@ -1,0 +1,5 @@
+import { Divider } from "./Divider";
+
+export { Divider } from "./Divider";
+export type { DividerProps } from "./divider.types";
+export default Divider;