+++ packages/implementations/react/src/index.ts
@@ -13,3 +13,4 @@ export * from "@rarui-react/text";
 export * from "@rarui-react/title";
 export * from "@rarui-react/toast";
 export * from "@rarui-react/toggle";
+export * from "@rarui-react/divider";