+++ packages/styles/src/index.ts
@@ -17,10 +17,12 @@ export { title } from "./packages/atomic/title";
 export { banner } from "./packages/atomic/banner";
 export { toast } from "./packages/atomic/toast";
 export { toggle } from "./packages/atomic/toggle";
+export { divider } from "./packages/atomic/divider";
 
 // types
 export type { BoxSprinkle } from "./packages/atomic/box";
 export type { ChipSprinkle } from "./packages/atomic/chip";
+export type { DividerSprinkle } from "./packages/atomic/divider";
 export type { IconSprinkle } from "./packages/atomic/icon";
 export type { TextSprinkle } from "./packages/atomic/text";
 export type { SkeletonSprinkle } from "./packages/atomic/skeleton";