+++ packages/styles/src/packages/atomic/divider/index.ts
@@ -1,0 +1,7 @@
+import { styles, sprinkle } from "./rarui-divider.css";
+
+export const divider = {
+  classnames: { ...styles },
+  sprinkle,
+};
+export type { DividerSprinkle } from "./rarui-divider.types";